+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `--verifier` flag to `blueprint verify` command to specify verifier ID
+
+### Changed
+
+- `blueprint verify` uses a new config structure that supports multiple verifiers
+
 ## [0.42.0] - 2025-12-04
 
 ### Added